subtests.docker_cli.kill: Split kill test into multiple ones
Split docker_cli/kill test into docker_cli/{kill,kill_bad,
kill_parallel_stress,kill_stopped,kill_stress} tests.

Additionally change the test_duration from 10s to 60s (for test
development 10s was used, 60s is the least what can be called testing)
