+++ config_defaults/subtests/docker_cli/kill_bad.ini
@@ -1,0 +1,12 @@
+[docker_cli/kill_bad]
+#: how long to wait before using the container
+wait_start = 1
+#: modifies the ``docker run`` options
+run_options_csv = -i
+subsubtests = bad
+#: modifies the container command
+exec_cmd = 'while :; do sleep 0.1; done'
+
+[docker_cli/kill_bad/bad]
+#: csv of bad signals which will be used in test
+bad_signals = 0,-1,-78,48,96,SIGBADSIGNAL,SIG,%%,!,\\,\,'',"", ,
