+++ config_defaults/subtests/docker_cli/kill_stopped.ini
@@ -1,0 +1,36 @@
+[docker_cli/kill_stopped]
+docker_timeout = 60
+#: how long to wait before using the container
+wait_start = 3
+#: modifies the ``docker run`` options
+run_options_csv = --attach=stdout
+subsubtests = sigstop,sigstop_ttyoff,sigstop_sigproxy,sigstop_sigproxy_ttyoff
+#: which signals should not be used (uncatchable signals)
+skip_signals = 9 17 19 27
+#: checking output produced by signal
+check_stdout = Received %%s, ignoring...
+#: range of used signals
+kill_signals = 1 31
+#: map signal numbers to names (RANDOM,true,false)
+kill_map_signals = false
+#: maximal acceptable delay caused by stress command
+stress_cmd_timeout = 5
+#: modifies the container command
+exec_cmd = 'for NUM in `seq 1 64`; do trap "echo Received $NUM, ignoring..." $NUM; done; while :; do sleep 0.1; done'
+#: allows you to force given sequence of signals
+signals_sequence =
+#: used kill command (``false`` -> ``docker kill $name``;
+#: `true`` -> ``os.kill $docker_cmd.pid``
+kill_sigproxy =
+#: execute detacched container and attach it in separate process
+run_container_attached = false
+
+[docker_cli/kill_stopped/sigstop]
+
+[docker_cli/kill_stopped/sigstop_ttyoff]
+
+[docker_cli/kill_stopped/sigstop_proxy]
+kill_sigproxy = true
+
+[docker_cli/kill_stopped/sigstop_proxy_ttyoff]
+kill_sigproxy = true
